Partial fix for ENG-511. Lacks a test case and there is still the theoretical potential for a buggy serializable exception to throw another exception while serializing and this would be uncaught.

+++ src/frontend/org/voltdb/ClientResponseImpl.java
@@ -109,7 +109,7 @@ public class ClientResponseImpl implements FastSerializable, ClientResponse {
     @Override
     public void readExternal(FastDeserializer in) throws IOException {
         in.readByte();//Skip version byte
-        status = (byte) in.readByte();
+        status = in.readByte();
         clusterRoundTripTime = in.readInt();
         m_exception = SerializableException.deserializeFromBuffer(in.buffer());
         results = (VoltTable[]) in.readArray(VoltTable.class);
@@ -129,7 +129,7 @@ public class ClientResponseImpl implements FastSerializable, ClientResponse {
             m_exception.serializeToBuffer(b);
             out.write(b.array());
         } else {
-            out.writeShort(0);//Length zero exception
+            out.writeInt(0);//Length zero exception
         }
         out.writeArray(results);
         out.writeString(extra);
